requirements.txt
# Python dependencies for Telemorph-Prime testing
kafka-python==2.0.2
